Purge removed tracks after 10x the lost track time
Removed tracks were being kept for as long as the particular instance
of BYTETracker was alive. This will cause memory to grow indefinitely
and is especially noticeable in busy scenes like common areas.

This seems safe as the removed tracks array is used to filter the
lost tracks array for tracks that are past the lost track time.

Added a comment to note that the original code would still report
the removed tracks as lost tracks at least once.